+++ pkg/rule/builtin/builtin.go
@@ -7,4 +7,5 @@ func Register(reg *rule.Registry) {
 	reg.Register(DetailsBlankLine{})
 	reg.Register(TableColumnCount{})
 	reg.Register(NoMissingSpaceATX{})
+	reg.Register(HeadingStartLeft{})
 }